Cleanup, add const, override, extern, move, etc. where appropriate Fix clang-tidy warnings Update comments

+++ common/WhirlyGlobeLib/include/BasicDrawableInstance.h
@@ -109,7 +109,7 @@ public:
     void setDrawPriority(int newPriority);
     
     /// Set the line width
-    void setLineWidth(int newLineWidth);
+    void setLineWidth(float newLineWidth);
     
     // Time we start counting from for motion
     void setStartTime(TimeInterval inStartTime);
